Add convenience Spawn Job
